chore(tag): move tag managers to own file
